Restructure of the code
