+++ app/src/main/java/com/github/se/gatherspot/model/event/EventStatus.kt
@@ -1,0 +1,11 @@
+package com.github.se.gatherspot.model.event
+
+/**
+ * Enum class for the status of an event
+ */
+enum class EventStatus {
+    DRAFT,
+    CREATED,
+    ON_GOING,
+    COMPLETED
+}
